Moved parts of recomp_port.h into new internal headers in src folder
